requirements.txt
flask
spacy
pdfplumber
python-docx
pandas
scikit-learn
